requirements.txt
py==1.4.31
pytest==2.8.7
scapy-python3==0.18
netaddr=0.7.18
PyYAML=3.11